install: Many fixes and improvements in installation scripts

+++ install/install_basic_dependencies.sh
@@ -1,20 +1,22 @@
 #!/bin/sh
 # Install all packages needed to run NERD and run all the services
 
+# disable "fastestmirror plugin, which in facat slows down yum"
+alias yum="yum --disableplugin=fastestmirror"
+
 echo "=============== Install basic dependencies ==============="
 
 echo "** Installing basic RPM packages **"
-#yum install -y https://centos7.iuscommunity.org/ius-release.rpm
-yum install -y epel-release
-yum install -y git wget gcc vim python36 python36-devel python36-setuptools python-setuptools
+yum install -y -q epel-release
+yum install -y -q git wget gcc vim python36 python36-devel python36-setuptools python-setuptools
 
 echo "** Installing pip and Python packages **"
 easy_install-2.7 --prefix /usr pip # Py2 is needed for Supervisor (until stable Supervisor 4 is out, which should work under Py3)
 easy_install-3.6 --prefix /usr pip
 # for some reason, this creates file /usr/bin/pip3.7 instead of pip3.6 (but everything works OK)
 
-# Allow to run python3.6 as python3
-ln -s /usr/bin/python3.6 /usr/bin/python3
+# Allow to run python3.6 as python3 (not needed, is created automatically)
+# ln -s /usr/bin/python3.6 /usr/bin/python3
 
 pip3 install -r /tmp/nerd_install/pip_requirements_nerdd.txt
 pip3 install -r /tmp/nerd_install/pip_requirements_nerdweb.txt
@@ -48,7 +50,7 @@ enabled=1
 gpgkey=https://www.mongodb.org/static/pgp/server-3.6.asc
 ' > /etc/yum.repos.d/mongodb-org-3.6.repo
 
-yum install -y mongodb-org
+yum install -y -q mongodb-org
 
 # ** Set up logrotate **
 # Configure Mongod to only reopen file after receiving SIGUSR1
@@ -77,7 +79,7 @@ systemctl start mongod.service
 
 
 echo "** Installing Redis **"
-yum install -y redis
+yum install -y -q redis
 
 echo "** Starting Redis **"
 systemctl enable redis.service
@@ -89,39 +91,62 @@ echo "** Installing RabbitMQ **"
 # We need more recent version than in CentOS7 (>=3.7.0), so install from developer sites
 
 # Install Erlang (dependency)
-echo '[rabbitmq-erlang]
-name=rabbitmq-erlang
-baseurl=https://dl.bintray.com/rabbitmq/rpm/erlang/20/el/7
-gpgcheck=1
-gpgkey=https://dl.bintray.com/rabbitmq/Keys/rabbitmq-release-signing-key.asc
-repo_gpgcheck=0
+echo '[rabbitmq_erlang]
+name=rabbitmq_erlang
+baseurl=https://packagecloud.io/rabbitmq/erlang/el/6/$basearch
+repo_gpgcheck=1
+gpgcheck=0
 enabled=1
-' > /etc/yum.repos.d/rabbitmq-erlang.repo
+gpgkey=https://packagecloud.io/rabbitmq/erlang/gpgkey
+sslverify=1
+sslcacert=/etc/pki/tls/certs/ca-bundle.crt
+metadata_expire=300
 
-yum install -y erlang
+[rabbitmq_erlang-source]
+name=rabbitmq_erlang-source
+baseurl=https://packagecloud.io/rabbitmq/erlang/el/6/SRPMS
+repo_gpgcheck=1
+gpgcheck=0
+enabled=1
+gpgkey=https://packagecloud.io/rabbitmq/erlang/gpgkey
+sslverify=1
+sslcacert=/etc/pki/tls/certs/ca-bundle.crt
+metadata_expire=300
+' > /etc/yum.repos.d/rabbitmq_erlang.repo
+
+yum install -y -q erlang
 
 # Install RabbitMQ
-yum install -y https://github.com/rabbitmq/rabbitmq-server/releases/download/v3.7.6/rabbitmq-server-3.7.6-1.el7.noarch.rpm
+if ! yum list installed rabbitmq-server >/dev/null 2>&1 ; then
+  yum install -y -q https://github.com/rabbitmq/rabbitmq-server/releases/download/v3.7.15/rabbitmq-server-3.7.15-1.el7.noarch.rpm
+fi
+
+# For some reason, hostname in Vagrant is often not set to anything meaningful,
+# but erlang needs to know it - this helps. 
+# Reference: https://stackoverflow.com/questions/45425286/rabbitmq-server-dont-start-unable-to-connect-to-epmd-ubuntu-16-04 
+# echo "HOSTNAME=localhost" >/etc/rabbitmq/rabbitmq-env.conf
 
 # Allow guest user to login remotely (allowed only from localhost by default)
 # This is necessary for Vagrant, but DON'T DO THIS IN PRODUCTION!
 # (rabbitmq.conf is not present after installation, so we just create it)
 if [ -d /vagrant ] ; then
   echo "It seems we run in Vagrant, allowing RabbitMQ guest user to login remotely."
-  echo "loopback_users = none" > /etc/rabbitmq/rabbitmq.conf
+  echo -e "loopback_users = none" > /etc/rabbitmq/rabbitmq.conf
 fi
+
+# Enable necessary plugins
+rabbitmq-plugins enable rabbitmq_management
+# rabbitmq-plugins enable rabbitmq_consistent_hash_exchange
 
 echo "** Starting RabbitMQ **"
 systemctl enable rabbitmq-server
 systemctl start rabbitmq-server
 
-# Enable necessary plugins
-rabbitmq-plugins enable rabbitmq_management
-rabbitmq-plugins enable rabbitmq_consistent_hash_exchange
-
 # Get rabbitmqadmin tool (provided via local API by the management plugin)
-wget -q http://localhost:15672/cli/rabbitmqadmin -O /usr/bin/rabbitmqadmin
-chmod +x /usr/bin/rabbitmqadmin
+if ! [ -f /usr/bin/rabbitmqadmin ] ; then
+  wget -q http://localhost:15672/cli/rabbitmqadmin -O /usr/bin/rabbitmqadmin
+  chmod +x /usr/bin/rabbitmqadmin
+fi
 
 
 
@@ -131,8 +156,10 @@ pip2 install supervisor
 
 
 echo "** Installing PostgreSQL **"
-yum install -y https://download.postgresql.org/pub/repos/yum/11/redhat/rhel-7-x86_64/pgdg-centos11-11-2.noarch.rpm
-yum install -y postgresql11-server postgresql11-devel
+if ! yum list installed postgresql11-server >/dev/null 2>&1 ; then
+  yum install -y -q https://download.postgresql.org/pub/repos/yum/11/redhat/rhel-7-x86_64/pgdg-centos11-11-2.noarch.rpm
+  yum install -y -q postgresql11-server postgresql11-devel
+fi
 
 # Initialize database (creates DB files in /var/lib/pgsql/11/data/)
 if ! [ -e /var/lib/pgsql/11/data/PG_VERSION ] ; then
